# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Dating Project",
        "description": "This is a sample Dating API server.",
        "version": "1.0.0"
    },
    "paths": {
        "/api/auth/signup": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "operationId": "App\\Http\\Controllers\\Api\\AuthController::signup",
                "requestBody": {
                    "description": "Create user object",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "firstname": {
                                        "type": "string"
                                    },
                                    "middlename": {
                                        "type": "string"
                                    },
                                    "lastname": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email"
                                    },
                                    "password": {
                                        "type": "string"
                                    },
                                    "password_confirmation": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "User registered",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/auth/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "operationId": "App\\Http\\Controllers\\Api\\AuthController::login",
                "requestBody": {
                    "description": "Returns API token with given user email and password",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email"
                                    },
                                    "password": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "User sign in",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "access_token": {
                                            "type": "string"
                                        },
                                        "token_type": {
                                            "type": "string"
                                        },
                                        "expires_at": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/auth/logout": {
            "get": {
                "tags": [
                    "Auth"
                ],
                "operationId": "App\\Http\\Controllers\\Api\\AuthController::logout",
                "requestBody": {
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "User logout"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "UserFilter": {
                "type": "object",
                "allOf": [
                    {
                        "properties": {
                            "id": {
                                "type": "integer",
                                "format": "integer"
                            },
                            "reason": {
                                "type": "object",
                                "allOf": [
                                    {
                                        "$ref": "#/components/schemas/UserSearchReason"
                                    }
                                ]
                            },
                            "gender": {
                                "type": "string"
                            },
                            "age_from": {
                                "type": "integer"
                            },
                            "age_to": {
                                "type": "integer"
                            },
                            "city_id": {
                                "type": "integer"
                            }
                        },
                        "type": "object"
                    }
                ]
            },
            "UserInfoOption": {
                "type": "object",
                "allOf": [
                    {
                        "properties": {
                            "id": {
                                "type": "integer",
                                "format": "integer"
                            },
                            "name": {
                                "type": "string",
                                "format": "string"
                            }
                        },
                        "type": "object"
                    }
                ]
            },
            "UserInformation": {
                "type": "object",
                "allOf": [
                    {
                        "properties": {
                            "relationship": {
                                "type": "object",
                                "allOf": [
                                    {
                                        "$ref": "#/components/schemas/UserInfoOption"
                                    }
                                ]
                            },
                            "living": {
                                "type": "object",
                                "allOf": [
                                    {
                                        "$ref": "#/components/schemas/UserInfoOption"
                                    }
                                ]
                            },
                            "children": {
                                "type": "object",
                                "allOf": [
                                    {
                                        "$ref": "#/components/schemas/UserInfoOption"
                                    }
                                ]
                            },
                            "smoking": {
                                "type": "object",
                                "allOf": [
                                    {
                                        "$ref": "#/components/schemas/UserInfoOption"
                                    }
                                ]
                            },
                            "drinking": {
                                "type": "object",
                                "allOf": [
                                    {
                                        "$ref": "#/components/schemas/UserInfoOption"
                                    }
                                ]
                            }
                        },
                        "type": "object"
                    }
                ]
            },
            "UserOptions": {
                "type": "object",
                "allOf": [
                    {
                        "properties": {
                            "gender": {
                                "type": "string",
                                "format": "string"
                            },
                            "birthday": {
                                "type": "string",
                                "format": "date"
                            },
                            "country_id": {
                                "type": "integer",
                                "format": "integer"
                            },
                            "region_id": {
                                "type": "integer",
                                "format": "integer"
                            },
                            "city_id": {
                                "type": "integer",
                                "format": "integer"
                            }
                        },
                        "type": "object"
                    }
                ]
            },
            "UserSearchReason": {
                "type": "object",
                "allOf": [
                    {
                        "properties": {
                            "id": {
                                "type": "integer",
                                "format": "integer"
                            },
                            "name": {
                                "type": "string",
                                "format": "string"
                            }
                        },
                        "type": "object"
                    }
                ]
            },
            "User": {
                "type": "object",
                "allOf": [
                    {
                        "properties": {
                            "id": {
                                "type": "integer",
                                "format": "integer"
                            },
                            "middlename": {
                                "type": "string",
                                "format": "string"
                            },
                            "lastname": {
                                "type": "string",
                                "format": "string"
                            },
                            "avatar_url": {
                                "type": "string",
                                "format": "string"
                            },
                            "last_activity": {
                                "type": "string",
                                "format": "date-time"
                            },
                            "options": {
                                "type": "object",
                                "allOf": [
                                    {
                                        "$ref": "#/components/schemas/UserOptions"
                                    }
                                ]
                            },
                            "information": {
                                "type": "object",
                                "allOf": [
                                    {
                                        "$ref": "#/components/schemas/UserInformation"
                                    }
                                ]
                            },
                            "filters": {
                                "type": "object",
                                "allOf": [
                                    {
                                        "$ref": "#/components/schemas/UserFilter"
                                    }
                                ]
                            }
                        },
                        "type": "object"
                    }
                ]
            },
...
        }
    }
}